requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.118.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jiter==0.11.0
openai==2.1.0
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
pydantic==2.11.9
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
sentence-transformers==2.2.2
torch==2.0.1
numpy==1.24.3